requirements.txt
streamlit
pinecone-client==3.1.0
PyPDF2 
pdfplumber
cohere 
transformers
numpy